Moved directional wspd calculation and recalculated wdir after resampling (#96, #97)
## Fixing #97
Directional wind speeds were previously calculated in L0toL1. In case of malfunctioning anemometer, wspd/wdir would be flagged and removed in the L1toL2 step, but not wspd_x/wspd_y, unless specified in the flag.csv file. Additionally, since wind direction needs to be re-calculated after time averages, then wrong periods would re-appear in wdir in those average files.

## Fixing #96
Due to a misunderstanding of the IDL code, the re-calculation of wdir from wspd_x/wpsd_y was  previously done in L2toL3. In the best case it would have no effect, since wspd_x/wpsd_y were previously calculated from wdir/wspd in L0toL1. In the worst case, flagged data in wdir, but not is wspd_x/wpsd_y (see above) would be re-inserted in the data. The re-calculation of wdir was removed from L2toL3 and added as part of AWS.py in the resampleL3 function.

+++ src/pypromice/process/L1toL2.py
@@ -27,9 +27,9 @@ def toL2(L1, T_0=273.15, ews=1013.246, ei0=6.1071, eps_overcast=1.,
         default is 6.1071.
     eps_overcast : int, optional
         Cloud overcast. The default is 1..
-    eps_clear : int, optional
+    eps_clear : float, optional
         Cloud clear. The default is 9.36508e-6.
-    emissivity : int, optional
+    emissivity : float, optional
         Emissivity. The default is 0.97.
 
     Returns
